chore: change base classes to be interfaces and rework the directory structure to have separate files for each class and subdirectories to reduce clutter
