+++ nodejs/src/recording-api/cache/index.ts
@@ -1,0 +1,2 @@
+export { MemoryCachedKeyStore } from './memory-cache'
+export { RedisCachedKeyStore } from './redis-cache'
