+++ nodejs/src/recording-api/crypto/tsconfig.json
@@ -1,0 +1,3 @@
+{
+    "extends": "../../../tsconfig.json"
+}
