+++ nodejs/src/recording-api/keystore/dynamodb-keystore.test.ts
@@ -1,19 +1,13 @@
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb'
 import { KMSClient } from '@aws-sdk/client-kms'
 
-import { RetentionService } from '../session-recording/retention/retention-service'
-import { TeamService } from '../session-recording/teams/team-service'
-import * as envUtils from '../utils/env-utils'
-import { InMemoryKeyStore, KeyStore, PassthroughKeyStore, getKeyStore } from './keystore'
-import { SessionKeyDeletedError } from './types'
+import { RetentionService } from '../../session-recording/retention/retention-service'
+import { TeamService } from '../../session-recording/teams/team-service'
+import { SessionKeyDeletedError } from '../types'
+import { DynamoDBKeyStore } from './dynamodb-keystore'
 
-jest.mock('../utils/env-utils', () => ({
-    ...jest.requireActual('../utils/env-utils'),
-    isCloud: jest.fn(),
-}))
-
-describe('KeyStore', () => {
-    let keyStore: KeyStore
+describe('DynamoDBKeyStore', () => {
+    let keyStore: DynamoDBKeyStore
     let mockDynamoDBClient: jest.Mocked<DynamoDBClient>
     let mockKMSClient: jest.Mocked<KMSClient>
     let mockRetentionService: jest.Mocked<RetentionService>
@@ -23,7 +17,7 @@ describe('KeyStore', () => {
     const mockEncryptedKey = new Uint8Array([101, 102, 103, 104, 105])
 
     describe('create and start', () => {
-        it('should create a KeyStore instance and initialize sodium on start', async () => {
+        it('should create a DynamoDBKeyStore instance and initialize sodium on start', async () => {
             const dynamoDBClient = { send: jest.fn() } as unknown as jest.Mocked<DynamoDBClient>
             const kmsClient = { send: jest.fn() } as unknown as jest.Mocked<KMSClient>
             const retentionService = {
@@ -33,10 +27,10 @@ describe('KeyStore', () => {
                 getEncryptionEnabledByTeamId: jest.fn().mockResolvedValue(true),
             } as unknown as jest.Mocked<TeamService>
 
-            const store = new KeyStore(dynamoDBClient, kmsClient, retentionService, teamService)
+            const store = new DynamoDBKeyStore(dynamoDBClient, kmsClient, retentionService, teamService)
             await store.start()
 
-            expect(store).toBeInstanceOf(KeyStore)
+            expect(store).toBeInstanceOf(DynamoDBKeyStore)
         })
     })
 
@@ -63,7 +57,7 @@ describe('KeyStore', () => {
             getEncryptionEnabledByTeamId: jest.fn().mockResolvedValue(true),
         } as unknown as jest.Mocked<TeamService>
 
-        keyStore = new KeyStore(mockDynamoDBClient, mockKMSClient, mockRetentionService, mockTeamService)
+        keyStore = new DynamoDBKeyStore(mockDynamoDBClient, mockKMSClient, mockRetentionService, mockTeamService)
         await keyStore.start()
     })
 
@@ -402,254 +396,6 @@ describe('KeyStore', () => {
             keyStore.stop()
 
             expect(mockDestroy).toHaveBeenCalledTimes(2)
-        })
-    })
-})
-
-describe('PassthroughKeyStore', () => {
-    let keyStore: PassthroughKeyStore
-
-    beforeEach(() => {
-        keyStore = new PassthroughKeyStore()
-    })
-
-    describe('start', () => {
-        it('should complete without error', async () => {
-            await expect(keyStore.start()).resolves.toBeUndefined()
-        })
-    })
-
-    describe('generateKey', () => {
-        it('should return empty keys with sessionState cleartext', async () => {
-            const result = await keyStore.generateKey('session-123', 1)
-
-            expect(result.plaintextKey).toEqual(Buffer.alloc(0))
-            expect(result.encryptedKey).toEqual(Buffer.alloc(0))
-            expect(result.sessionState).toBe('cleartext')
-        })
-    })
-
-    describe('getKey', () => {
-        it('should return empty keys with sessionState cleartext', async () => {
-            const result = await keyStore.getKey('session-123', 1)
-
-            expect(result.plaintextKey).toEqual(Buffer.alloc(0))
-            expect(result.encryptedKey).toEqual(Buffer.alloc(0))
-            expect(result.sessionState).toBe('cleartext')
-        })
-    })
-
-    describe('deleteKey', () => {
-        it('should return true', async () => {
-            const result = await keyStore.deleteKey('session-123', 1)
-
-            expect(result).toBe(true)
-        })
-    })
-
-    describe('stop', () => {
-        it('should complete without error', () => {
-            expect(() => keyStore.stop()).not.toThrow()
-        })
-    })
-})
-
-describe('getKeyStore', () => {
-    let mockTeamService: jest.Mocked<TeamService>
-    let mockRetentionService: jest.Mocked<RetentionService>
-
-    beforeEach(() => {
-        mockTeamService = {
-            getEncryptionEnabledByTeamId: jest.fn().mockResolvedValue(true),
-        } as unknown as jest.Mocked<TeamService>
-
-        mockRetentionService = {
-            getSessionRetentionDays: jest.fn().mockResolvedValue(30),
-        } as unknown as jest.Mocked<RetentionService>
-    })
-
-    it('should return KeyStore when running on cloud', () => {
-        ;(envUtils.isCloud as jest.Mock).mockReturnValue(true)
-
-        const keyStore = getKeyStore(mockTeamService, mockRetentionService, 'us-east-1')
-
-        expect(keyStore).toBeInstanceOf(KeyStore)
-    })
-
-    it('should return PassthroughKeyStore when not running on cloud', () => {
-        ;(envUtils.isCloud as jest.Mock).mockReturnValue(false)
-
-        const keyStore = getKeyStore(mockTeamService, mockRetentionService, 'us-east-1')
-
-        expect(keyStore).toBeInstanceOf(PassthroughKeyStore)
-    })
-
-    it('should accept custom kmsEndpoint and dynamoDBEndpoint', () => {
-        ;(envUtils.isCloud as jest.Mock).mockReturnValue(true)
-
-        const keyStore = getKeyStore(mockTeamService, mockRetentionService, 'us-east-1', {
-            kmsEndpoint: 'http://localhost:4566',
-            dynamoDBEndpoint: 'http://localhost:4566',
-        })
-
-        expect(keyStore).toBeInstanceOf(KeyStore)
-    })
-})
-
-describe('InMemoryKeyStore', () => {
-    let keyStore: InMemoryKeyStore
-
-    beforeEach(async () => {
-        keyStore = new InMemoryKeyStore()
-        await keyStore.start()
-    })
-
-    describe('start', () => {
-        it('should initialize sodium', async () => {
-            const store = new InMemoryKeyStore()
-            await expect(store.start()).resolves.toBeUndefined()
-        })
-    })
-
-    describe('generateKey', () => {
-        it('should generate a key with ciphertext sessionState', async () => {
-            const result = await keyStore.generateKey('session-123', 1)
-
-            expect(result.sessionState).toBe('ciphertext')
-            expect(result.plaintextKey).toBeInstanceOf(Buffer)
-            expect(result.plaintextKey.length).toBe(32) // crypto_secretbox_KEYBYTES
-            expect(result.encryptedKey).toEqual(result.plaintextKey)
-        })
-
-        it('should generate different keys for different sessions', async () => {
-            const key1 = await keyStore.generateKey('session-1', 1)
-            const key2 = await keyStore.generateKey('session-2', 1)
-
-            expect(key1.plaintextKey.equals(key2.plaintextKey)).toBe(false)
-        })
-
-        it('should generate different keys for different teams with same session id', async () => {
-            const key1 = await keyStore.generateKey('session-1', 1)
-            const key2 = await keyStore.generateKey('session-1', 2)
-
-            expect(key1.plaintextKey.equals(key2.plaintextKey)).toBe(false)
-        })
-
-        it('should overwrite existing key when called again', async () => {
-            const key1 = await keyStore.generateKey('session-1', 1)
-            const key2 = await keyStore.generateKey('session-1', 1)
-
-            expect(key1.plaintextKey.equals(key2.plaintextKey)).toBe(false)
-
-            // Verify getKey returns the new key
-            const retrieved = await keyStore.getKey('session-1', 1)
-            expect(retrieved.plaintextKey.equals(key2.plaintextKey)).toBe(true)
-        })
-    })
-
-    describe('getKey', () => {
-        it('should return existing key if previously generated', async () => {
-            const generated = await keyStore.generateKey('session-123', 1)
-            const retrieved = await keyStore.getKey('session-123', 1)
-
-            expect(retrieved.plaintextKey.equals(generated.plaintextKey)).toBe(true)
-            expect(retrieved.encryptedKey.equals(generated.encryptedKey)).toBe(true)
-            expect(retrieved.sessionState).toBe('ciphertext')
-        })
-
-        it('should generate new key if not found', async () => {
-            const result = await keyStore.getKey('new-session', 1)
-
-            expect(result.sessionState).toBe('ciphertext')
-            expect(result.plaintextKey.length).toBe(32)
-        })
-
-        it('should return same key on subsequent calls', async () => {
-            const first = await keyStore.getKey('session-123', 1)
-            const second = await keyStore.getKey('session-123', 1)
-
-            expect(first.plaintextKey.equals(second.plaintextKey)).toBe(true)
-        })
-
-        it('should throw SessionKeyDeletedError if key was deleted', async () => {
-            await keyStore.generateKey('session-123', 1)
-            await keyStore.deleteKey('session-123', 1)
-
-            await expect(keyStore.getKey('session-123', 1)).rejects.toThrow(SessionKeyDeletedError)
-        })
-
-        it('should include deletedAt timestamp in SessionKeyDeletedError', async () => {
-            await keyStore.generateKey('session-123', 1)
-
-            const beforeDelete = Date.now()
-            await keyStore.deleteKey('session-123', 1)
-            const afterDelete = Date.now()
-
-            try {
-                await keyStore.getKey('session-123', 1)
-                fail('Expected SessionKeyDeletedError')
-            } catch (error) {
-                expect(error).toBeInstanceOf(SessionKeyDeletedError)
-                const deletedError = error as SessionKeyDeletedError
-                expect(deletedError.deletedAt).toBeGreaterThanOrEqual(beforeDelete)
-                expect(deletedError.deletedAt).toBeLessThanOrEqual(afterDelete)
-            }
-        })
-    })
-
-    describe('deleteKey', () => {
-        it('should return true if key existed', async () => {
-            await keyStore.generateKey('session-123', 1)
-            const result = await keyStore.deleteKey('session-123', 1)
-
-            expect(result).toBe(true)
-        })
-
-        it('should return false if key did not exist', async () => {
-            const result = await keyStore.deleteKey('non-existent', 999)
-
-            expect(result).toBe(false)
-        })
-
-        it('should prevent subsequent getKey calls', async () => {
-            await keyStore.generateKey('session-123', 1)
-            await keyStore.deleteKey('session-123', 1)
-
-            await expect(keyStore.getKey('session-123', 1)).rejects.toThrow(SessionKeyDeletedError)
-        })
-
-        it('should not affect other sessions', async () => {
-            await keyStore.generateKey('session-1', 1)
-            await keyStore.generateKey('session-2', 1)
-
-            await keyStore.deleteKey('session-1', 1)
-
-            // session-2 should still be accessible
-            const result = await keyStore.getKey('session-2', 1)
-            expect(result.sessionState).toBe('ciphertext')
-
-            // session-1 should throw
-            await expect(keyStore.getKey('session-1', 1)).rejects.toThrow(SessionKeyDeletedError)
-        })
-
-        it('should not affect same session id in different teams', async () => {
-            await keyStore.generateKey('session-1', 1)
-            await keyStore.generateKey('session-1', 2)
-
-            await keyStore.deleteKey('session-1', 1)
-
-            // Team 2 should still have access
-            const result = await keyStore.getKey('session-1', 2)
-            expect(result.sessionState).toBe('ciphertext')
-
-            // Team 1 should throw
-            await expect(keyStore.getKey('session-1', 1)).rejects.toThrow(SessionKeyDeletedError)
-        })
-    })
-
-    describe('stop', () => {
-        it('should complete without error', () => {
-            expect(() => keyStore.stop()).not.toThrow()
         })
     })
 })
